Disable retry on cleanup logout
It may occur that the exit logout could get hung
up for one reason or another. Change to opportunistically
trying to logout.

It would have to fail generally around 3 times within a minute
for it to be a problem, so it is unlikely to ever
be a problem should it fail.

Change-Id: Ie32b2ae03618fe078e3965d9ee68dda962206bfe