Slow down setting configuration
Problems may occur if setting configuration in
a script and multiple transactions are attempted
back to back.

Also, it can be confusing to set settings and then readback the old settings.

Address both concerns by only returning from
set when the transaction manifests in the data
returned from the peer system.

Change-Id: I93f360e0e2a3360bed460534bb924df1472d2bfb