Retry login on dead session
Sometimes a BMC just up and drops an IPMI session without warning.

The only way to recover is to login, so incorporate a login into the
retry logic.

Now a couple of retries are done the 'cheap' way (retransmit) followed
by a full login attempt.

This does mean that SOL sessions are less likely to survive a shared
nic outage during PXE boot or similar,
however we just have to take that hit.

Change-Id: Ica11d29519b1c69cbd151bce9100f8a7517544ca