Go back to trusting XCC health assessment
The health assesment can be much cheaper if the XCC health
assessment is trusted.

The issue would be that if the XCC web shows healthy, but the error LED
is lit, this condition would no longer be seen.

However, the error LED shouldn't be lit unless the XCC detects a fault.
So if such a condition of LED lit without explanation should arise in
the future, it would be better treated as an XCC firmware issue.

Change-Id: I55a4451eda4aba43fcf3d3e8e2307c3ce30b8fcd