Revert assertion mask usage
Unfortunately, different devices apply this
in different ways. The impetus was a BMC that
was incorrectly returning a bad reading, and a
thought toward using the assertion events instead.

This has two problems, one is that this mechanism is optional, and for another
the caller is expected to manage the arming.

Go back to using the sensor reading to let the BMC
mark on-the-fly the state.

Change-Id: Ie61d4b8427905171b32f6998a80806f1027660d6