Avoid doing extra AddNewVolWithNaAsync calls.
Using a trial and error method to determine the format of
AddNewVolWithNaAsync causes issues on some server models. When
repeatedly deleting and creating the RAID array on Cyborg with
CDI3A2U XCC firmware we occasionally see hanged web sessions and
on rare occasions an incorrectly formatted AddNewVolWithNaAsync
call will return success and not create the array.
To reduce the number of AddNewVolWithNaAsync calls we determine
when GetDefaultVolProp requires the extra controller slot number
argument and in this case we directly use the
AddNewVolWithNaAsync call with the extra controller type argument
as these 2 changes go hand in hand in the XCC firmware. This logic
reduces the amount of AddNewVolWithNaAsync calls by one and the
issues described above can no longer be observed.

Change-Id: Icdaa5ef0fb5629ad965e37cbbecdcf892f6e315f