Filter out DPU from systems collection if ambiguous
DPUs may manifest in a Systems collection.

If the caller does not specify sysurl, they are almost certainly wanting
to manage the 'main' system,
not an adapter.

Provide for this expectation by
retrying the collection with DPU
filtered out,
so that if there's exactly one non-DPU
system, then it will still be functional.

Change-Id: I300c47245111121440ae0f037d202b074407c45d